iv_fd: Allow timeout handling optimizations in poll methods.
All file descriptor polling methods that ivykis supports have timeout
arguments to their respective polling system calls -- epoll_wait(2),
poll(2) and poll(7d) have a relative milliseconds argument, and
kevent(2), port_getn(3C) and ppoll(2) have a relative struct timespec
pointer argument.

When calling into the kernel to wait for file descriptor activity to
occur, ivykis determines the timeout to be used for that system call
by finding the registered iv_timer with the soonest expiring timeout,
and passing in that timeout to the poll system call for the polling
method currently in use, so that that system call will return when
there is file descriptor activity or when the first iv_timer to be
expiring has expired.

One opportunity for improvement on this approach has to do with the
fact that on most operating systems, there appears to be a nontrivial
cost associated with setting up and tearing down a kernel timer on
every entry to and exit from a poll system call (which is what happens
if you pass in a sub-infinite timeout to such a system call), and that
some of those operating systems provide a way of setting up a kernel
timer that can be kept running across system calls.

Even on operating systems where setting up and tearing down kernel
timers isn't all that costly, it can be advantageous to set up a
separate kernel timer corresponding to the soonest expiring iv_timer,
as that will give us explicit notification of when that timer expires,
which can let us avoid calling clock_gettime() on every poll system
call return to check whether the current time is greater than or equal
to the soonest expiring iv_timer's timeout yet.

This commit adds two new poll method methods, ->set_poll_timeout()
and ->clear_poll_timeout():

* ->set_poll_timeout(), when implemented, will be called by the ivykis
  core when it wants to configure an explicit timeout, which when it
  expires must cause a subsequent or in-progress call to ->poll() to
  return (with a nonzero value), even if that call to ->poll() was done
  with a NULL (infinite) timeout.

* ->clear_poll_timeout() will be called if an explicit timeout was
  previously configured which has now been superseded, for example if
  an iv_timer with a timeout earlier than the currently registered
  explicit timeout is registered before the explicit timeout expires.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>