epoll: Implement ->set_poll_timeout() using timerfd_create(2).
Without this commit, iv_event_raw_bench produces the following
performance figures (median values of five runs) on a virtual machine
running CentOS 6.4 (which supports timerfd_create(2)):

	signal: 13762361 in 5000018382 nsec => 2752462/sec
	timer: 8747421 in 5000005528 nsec => 1749482/sec

This commit implements ->{set,clear}_poll_timeout() for the epoll poll
method using timerfd_create(2), allowing offloading of timeout handling
from epoll_wait(2) to a timerfd so that we can call epoll_wait(2) with
timeout values of -1 under the right conditions, which avoids setting
up and tearing down a kernel timer on every epoll_wait(2) call.  As a
side effect, this commit also allows avoiding calls to clock_gettime()
as long as no timers are expiring and noone is referencing iv_now or
calling iv_validate_now().  With these changes, the results are:

	timer: 13750406 in 5000013792 nsec => 2750073/sec

In other words, this commit virtually eliminates all of the cost of
having a background timer running.

To see which of the two effects (timer offloading or clock_gettime()
avoidance) helps the most, I modified this commit to always call
clock_gettime() in the ivykis main loop to check for expired timers
even if we know that no timers can have expired yet, and that gives us:

	timer: 10274371 in 5000013266 nsec => 2054868/sec

Which means that the timer offloading gives an improvement of about
300k loops per second, and the clock_gettime() avoidance gives an
additional improvement on top of that of about 700k loops per second.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>