port: Implement ->set_poll_timeout() using timer_create(3C).
Without this commit, iv_event_raw_bench produces the following
performance figures (median values of five runs) on a virtual machine
running Solaris 11.1:

	signal: 2584578 in 4996484168 nsec => 517279/sec
	timer: 2292693 in 5000012397 nsec => 458537/sec

This commit implements ->{set,clear}_poll_timeout() for the
port_create(3C) poll method using timer_create(3C), which allows
keeping a kernel timer running across port_getn(3C) calls, so that
we can call port_getn(3C) with NULL timeout values under the right
conditions, and it also allows avoiding calls to clock_gettime() as
long as no timers are expiring and noone is referencing iv_now or
calling iv_validate_now().  With these changes, the results are:

	timer: 2585753 in 5000023204 nsec => 517148/sec

So as for the corresponding epoll and kqueue changes, this commit
virtually eliminates all of the cost of having a background timer
running.

And as in the other two cases, I modified this commit to always call
clock_gettime() in the ivykis main loop to check for expired timers
even if we know that no timers can have expired yet, and that gives us:

	timer: 2309814 in 5000019122 nsec => 461961/sec

This means that the kernel timer setup and teardown overhead seems
minimal, but the savings obtained from being able to avoid doing a
lot of clock_gettime() calls are significant.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>